get basic all bundle to build
