+++ eslint.config.mjs
@@ -53,7 +53,7 @@ export default [
 			"@typescript-eslint/no-explicit-any": "off",
 			"@typescript-eslint/ban-ts-comment": "off",
 			"@typescript-eslint/ban-types": "off",
-
+			"@typescript-eslint/no-require-imports": "off",
 			"@typescript-eslint/no-unused-vars": [
 				"warn",
 				{
