+++ frontend/index.html
@@ -6,7 +6,7 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>Vite + TS</title>
 
-		<script src="/scram/scramjet.controller.js"></script>
+		<script src="/scram/scramjet.all.js"></script>
 	</head>
 
 	<body>
