+++ src/shared/index.ts
@@ -14,36 +14,26 @@ import { parseDomain } from "parse-domain";
 import { ScramjetHeaders } from "./headers";
 import { CookieStore } from "./cookie";
 import { htmlRules, unrewriteHtml } from "./rewriters/html";
-import { $scramjet } from "../scramjet";
+import { config } from "../shared";
+import { ScramjetFlags } from "../types";
 
-$scramjet.shared = {
-	util: {
-		parseDomain,
-		BareClient,
-		BareMuxConnection,
-		ScramjetHeaders,
-	},
-	url: {
-		rewriteUrl,
-		unrewriteUrl,
-		rewriteBlob,
-		unrewriteBlob,
-	},
-	rewrite: {
-		rewriteUrl,
-		rewriteCss,
-		unrewriteCss,
-		rewriteHtml,
-		unrewriteHtml,
-		rewriteSrcset,
-		rewriteJs,
-		rewriteHeaders,
-		rewriteWorkers,
-		htmlRules,
-	},
-	CookieStore,
-};
+export let codecEncode: (input: string) => string;
+export let codecDecode: (input: string) => string;
 
-if ("document" in self && document?.currentScript) {
-	document.currentScript.remove();
+const nativeFunction = Function;
+export function loadCodecs() {
+	codecEncode = nativeFunction(`return ${config.codec.encode}`)() as any;
+	codecDecode = nativeFunction(`return ${config.codec.decode}`)() as any;
+}
+
+export function flagEnabled(flag: keyof ScramjetFlags, url: URL): boolean {
+	const value = config.flags[flag];
+	for (const regex in config.siteFlags) {
+		const partialflags = config.siteFlags[regex];
+		if (new RegExp(regex).test(url.href) && flag in partialflags) {
+			return partialflags[flag];
+		}
+	}
+
+	return value;
 }
