+++ static/index.html
@@ -53,7 +53,7 @@
 		</style>
 		<script src="https://cdn.jsdelivr.net/npm/dreamland"></script>
 		<script src="/baremux/index.js"></script>
-		<script src="/scram/scramjet.controller.js"></script>
+		<script src="/scram/scramjet.all.js"></script>
 		<script src="config.js"></script>
 		<script src="store.js"></script>
 	</head>
